+++ CHANGELOG.md
@@ -10,6 +10,7 @@ SPDX-License-Identifier: GPL-3.0-or-later
     - Auto-move current box when typing (!259, !262).
     - Upgrade to JavaFx 25 (!261).
     - Update theme to AtlantaFx 2.1.0 (!263).
+    - Add dark theme (#66, !264).
 
 ## Version 0.11 - Not much but still - 2024/12/30