+++ CHANGELOG.md
@@ -3,7 +3,7 @@ SPDX-FileCopyrightText: 2025 Antoine Belvire
 SPDX-License-Identifier: GPL-3.0-or-later
 -->
 
-## Not released
+## Version 0.12 - Upgrade to Java 25 - 2025/11/09
 
 - Upgrade to Java 25 (!260).
 - Reduce memory usage (!267).